+++ test/libsolidity/syntaxTests/inheritance/override/public_constant_var_overrides_pure.sol
@@ -1,0 +1,7 @@
+abstract contract C {
+    function foo() external pure virtual returns (uint);
+}
+contract X is C {
+	uint public constant override foo = 7;
+}
+// ----