+++ Changelog.md
@@ -30,6 +30,7 @@ Compiler Features:
 Bugfixes:
  * NatSpec: Constructors and functions have consistent userdoc output.
  * Inheritance: Disallow public state variables overwriting ``pure`` functions.
+ * State Mutability: Constant public state variables are considered ``pure`` functions.
 
 
 ### 0.6.12 (unreleased)